==> notebooks/01_load_and_validate_raw_data.ipynb <==
import requests
from pathlib import Path

def download_one_file_of_raw_data(year: int, month: int) -> Path:
    """"""
    URL = f'https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet'
    response = requests.get(URL)

    if response.status_code == 200:
        path = f'../data/raw/yellow_tripdata_{year}-{month:02d}.parquet'
        open(path, 'wb').write(response.content)
        return Path(path)
    else:
        raise Exception(f'Failed to download {URL}')

download_one_file_of_raw_data(2023, 6)

import pandas as pd

rides = pd.read_parquet('../data/raw/yellow_tripdata_2023-06.parquet')
rides.head()

rides = rides[['tpep_pickup_datetime', 'PULocationID']]
rides.rename(columns={'tpep_pickup_datetime': 'pickup_datetime', 'PULocationID': 'pickup_location_id'}, inplace=True)

rides

rides['pickup_datetime'].describe()

rides = rides[rides['pickup_datetime'] < '2023-07-01']
rides = rides[rides['pickup_datetime'] >= '2023-06-01']
rides['pickup_datetime'].describe()

rides.to_parquet("../data/transformed/validated_rides_2023-06.parquet")
